use-cases/00-hello-world/python/pyproject.toml
[project]
name = "hello-world"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
]
